move general functions in chat/user data to separate abstract class
